use sre-wrapper container instead of the stuff in local-container folder
