+++ api/src/home/analysis-platform/data/AnalysisPlatformDS/Data/Analysis/exampleAnalysis.json
@@ -1,5 +1,5 @@
 {
-  "_id": "4483c9b0-d505-46c9-a157-94c79f4d7a6b",
+  "_id": "4483c9b0-d505-46c9-a157-94c79f4d7a6a",
   "type": "AnalysisPlatformDS/Blueprints/Analysis",
   "name": "exampleAnalysis",
   "description": "example analysis",
