--- local-container/docker-compose.yml
@@ -1,25 +1,0 @@
-version: "3.8"
-
-services:
-  wrapper:
-    build: .
-    image: wrapper
-#    environment:
-#      SIMA_LICENSE: |
-#
-#      PUBLIC_DMSS_API: http://localhost:5000
-    networks:
-      - data-modelling-storage-service_default
-    external_links:
-      - dmss
-      - db
-    volumes:
-      - ./wrapper.py:/code/wrapper.py
-      - ./init.sh:/code/init.sh
-    command: >-
-        /code/init.sh --result-reference-location=4fe42156-5c4b-414e-9dad-124c6f9f2ae9.jobs.0.result --token=XXXXXX --application-input='{"type":"AnalysisPlatformDS/Blueprints/SIMAApplicationInput","name":"example_app_input","description":"example application input for local container","inputType":"system/SIMOS/NamedEntity","outputType":"AnalysisPlatformDS/Blueprints/Analysis","input":"{\"type\":\"system/SIMOS/NamedEntity\",\"name\":\"exampleInputEntity\",\"description\":\"example\"}","applicationConfig":{"type":"AnalysisPlatformDS/Blueprints/SIMAApplicationConfig","description":"","name":"example_app_config","workflow":"someWorkflow","workflowTask":"someWorkflowTask","resultPath":"AnalysisPlatformDS/Data/Analysis/results"},"SIMAComputeConfig":{}}'
-
-
-networks:
-  data-modelling-storage-service_default:
-    external: true
