--- local-container/pyproject.toml
@@ -1,18 +1,0 @@
-[tool.poetry]
-name = "Job_Initscript"
-version = "0.1.0"
-description = ""
-authors = ["Your Name <you@example.com>"]
-license = "MIT"
-
-[tool.poetry.dependencies]
-python = "^3.8"
-click = "^8.0.4"
-dmss-api = "^0.3.17"
-
-
-[tool.poetry.dev-dependencies]
-
-[build-system]
-requires = ["poetry-core>=1.0.0"]
-build-backend = "poetry.core.masonry.api"
