+++ sre-wrapper/init.sh
@@ -22,6 +22,14 @@ for i in "$@"; do
       TOKEN="${i#*=}"
       shift # past argument=value
       ;;
+    --application-input=*)
+      APPLICATION_INPUT="${i#*=}"
+      shift # past argument=value
+      ;;
+    --result-reference-location=*)
+      RESULT_REFERENCE_LOCATION="${i#*=}"
+      shift # past argument=value
+      ;;
     --stask=*)
       STASK="${i#*=}"
       shift # past argument=value
@@ -46,14 +54,6 @@ for i in "$@"; do
       INPUT="${i#*=}"
       shift
       ;;
-    --target=*)
-      TARGET="${i#*=}"
-      shift
-      ;;
-    --result-link-target=*)
-      RESULT_LINK_TARGET="${i#*=}"
-      shift
-      ;;
     *)
       echo "WARNING: Invalid argument '$i'"
       ;;
@@ -61,15 +61,19 @@ for i in "$@"; do
 done
 
 
-# Run the DMT wrapper. Preparing the SRE environment
-/code/job_wrapper.py run --token=$TOKEN --stask=$STASK --task=$TASK --workflow=$WORKFLOW --compute-service-cfg=$COMPUTE_SERVICE_CFG $REMOTE_RUN --input=$INPUT
-# SIMA Headless
-/opt/sima/sre \
-  -data=$SRE_HOME \
-  -commands file=$SRE_HOME/commands.txt \
-  -consoleLog
-# Upload results
-  /code/job_wrapper.py upload --token=$TOKEN --target=$TARGET --result-link-target=$RESULT_LINK_TARGET --task=$TASK --workflow=$WORKFLOW
+## Run the DMT wrapper. Preparing the SRE environment
+#/code/job_wrapper.py run --token=$TOKEN --stask=$STASK --task=$TASK --workflow=$WORKFLOW --compute-service-cfg=$COMPUTE_SERVICE_CFG $REMOTE_RUN --input=$INPUT
+## SIMA Headless
+#/opt/sima/sre \
+#  -data=$SRE_HOME \
+#  -commands file=$SRE_HOME/commands.txt \
+#  -consoleLog
+## Upload results
+#  /code/job_wrapper.py upload --token=$TOKEN --target=$TARGET --result-link-target=$RESULT_LINK_TARGET --task=$TASK --workflow=$WORKFLOW
+APPLICATION_INPUT=`echo $APPLICATION_INPUT | sed 's/ /%20/g'` #replace space with %20 since click cannot have space in the option value
+
+/code/job_wrapper.py get-and-upload-result --token=$TOKEN --result-reference-location=$RESULT_REFERENCE_LOCATION --application-input=$APPLICATION_INPUT
+
 time=$(date +'%d/%m/%Y %r')
 echo "** End time: ${time} **"
